pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mags-codedev"
version = "0.1.0"
description = "A Multi-Agent LLM Developer Workflow using LangGraph"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "langgraph>=0.1.0",
    "langchain-core>=0.1.50",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "google-generativeai>=0.3.0",
    "langchain-anthropic>=0.1.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "gitpython>=3.1.40",
    "docker>=7.0.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0"
]

[project.scripts]
mags-codedev = "mags_codedev.cli:app"

[tool.setuptools.packages.find]
include = ["mags_codedev", "mags_codedev.*"]
